+++ src/server/services/task/index.ts
@@ -210,6 +210,7 @@ export class TaskService {
         return {
           author: task.assigneeAgentId ? authorMap.get(task.assigneeAgentId) : undefined,
           id: t.topicId ?? undefined,
+          runningOperation: t.metadata?.runningOperation ?? null,
           seq: t.seq,
           status: t.status,
           summary: handoff?.summary,
